ai-bos-platform/backend/pyproject.toml
[project]
name = "ai-bos-backend"
version = "0.1.0"
description = "AI-BOS Backend Orchestrator"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "crewai",
    "langchain",
    "langchain-community",
    "langchain-google-genai",
    "duckduckgo-search",
    "requests",
    "huggingface-hub>=1.6.0",
]

[tool.uv]
managed = true
